backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
pydantic==2.8.2
pydantic-settings==2.4.0
SQLAlchemy==2.0.34
psycopg[binary]==3.2.1
alembic==1.13.2
httpx==0.27.2
email-validator==2.2.0
pytest==8.3.2
